fix: cross-platform CI failures (arm32, arm64-qemu, macOS)
- Fix sign-compare error in test_io_flow_multiclient.c (time_t vs uint32_t)
- Add darwin/dap_io_flow_darwin_gcd.c to macOS build (missing symbols)
- Add platform guards for DAP_IO_FLOW_LB_TIER_CLASSIC_BPF in regression tests
- Skip GNU --wrap linker options on Apple (use DYLD_INSERT_LIBRARIES instead)
- Graceful skip for CBPF attach failures in qemu/emulated environments
- Reduce per-packet log noise in test_trans_helpers receive callback
- Increase CTest timeout for cross-compiled integration tests (120s→600s)
- Fix format warnings: %lu → PRIu64 for uint64_t on 32-bit platforms

Made-with: Cursor

+++ tests/integration/net/trans/test_trans_integration.c
@@ -17,6 +17,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <inttypes.h>
 #include <pthread.h>
 #include <unistd.h>
 #include <signal.h>
@@ -193,8 +194,8 @@ static void print_test_statistics(void)
     printf("📊 Overall Results:\n");
     printf("   Total Scenarios: %zu passed, %zu failed\n",
            s_test_stats.total_scenarios_passed, s_test_stats.total_scenarios_failed);
-    printf("   Total Duration: %lu seconds (%.1f minutes)\n",
-           total_duration_sec, total_duration_sec / 60.0);
+    printf("   Total Duration: %" PRIu64 " seconds (%.1f minutes)\n",
+           total_duration_sec, (double)total_duration_sec / 60.0);
     printf("\n");
     
     // Per-transport statistics
